+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,14 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.1] - Unreleased
+
+### Changed
+
+- The protection against exiting an interactive shell with suspended jobs, which
+  was introduced in 3.2.0, can now be disabled by setting the `posixlycorrect`
+  option.
+
 ## [3.2.0] - 2026-06-15
 
 ### Added
@@ -353,6 +361,7 @@ later.
 
 - Initial release of the shell
 
+[3.2.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.2.1
 [3.2.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.2.0
 [3.1.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.1.0
 [3.0.8]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.8